Move ray_utils.py from engine to executor package
Per @zhuohan123's request here: https://github.com/vllm-project/vllm/pull/3466#discussion_r1575575587
